requirements.txt
# MCP Server Requirements
# Core MCP and FastMCP dependencies
mcp>=1.8.0,<2.0.0
fastmcp>=2.3.4,<2.4.0

# Atlassian API
atlassian-python-api>=4.0.0
requests[socks]>=2.31.0

# HTTP and async
httpx>=0.28.0
uvicorn>=0.27.1
starlette>=0.37.1

# Data processing
beautifulsoup4>=4.12.3
markdownify>=0.11.6
markdown>=3.7.0
pydantic>=2.10.6

# Utilities
python-dotenv>=1.0.1
trio>=0.29.0
click>=8.1.7
thefuzz>=0.22.1
python-dateutil>=2.9.0.post0
types-python-dateutil>=2.9.0.20241206
keyring>=25.6.0
cachetools>=5.0.0
types-cachetools>=5.5.0.20240820

